pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai[typing,typings]==1.12.1",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]
